Update readDirStream to use Maybe
This patch changes `readDirStream` to signal end of directory with a
Nothing value. In addition it changes the wrapped readdir function to
only return -1 in the case of an actual error. This change allows the
errno handling logic to take advantage of helpers from Foreign.C.Error,
simplifying the logic.

Fixes #81